Add ability to load the dump using a tool/command
The db module now accepts a `dumptool` configuration item
where the user can enter a command to execute his/her prefered
tool or wrapper script.
The dsn is parsed into variables and are made available, with
the variables from the module configuration, to the command
for interpolation, allowing the executing of commands like:

`dumptool: 'mysql -u $user -p $password -h $host -D $dbname < $dump'`

The sqlite module has been cleaned up to respect the reconnect
settings and to aviod dependency on the order of the execution
of its parent method. Mainly because there is no need to "reload"
the connection when touching the sqlite db file...